ATEM fader handling
closes #5

+++ software/firmware/VideoCtrl/lib/hw/PCA9685.cpp
@@ -55,7 +55,7 @@ void PCA9685::setLEDDimmed(uint8_t ledNumber, uint8_t amount) {		// Amount from 
 		/*int randNumber = (int)random(4096);	// Randomize the phaseshift to distribute load. Good idea? Hope so. */
 		// uint16_t randNumber = 1000;
 	    // pseudo rand: ledNumber * 256 (max: 15 -> 15*256 = 3540)
-	    uint16_t randNumber = ledNumber << 8;
+	    uint16_t randNumber = ((uint16_t)ledNumber) << 8;
 		writeLED(ledNumber, randNumber, (uint16_t)(amount*41+randNumber) & 0xFFF);
 	}
 }
